+++ __docs__/wonder-blocks-switch/switch.stories.tsx
@@ -5,7 +5,7 @@ import {expect} from "@storybook/jest";
 import {userEvent, within} from "@storybook/testing-library";
 
 import Switch from "@khanacademy/wonder-blocks-switch";
-import {View} from "@khanacademy/wonder-blocks-core";
+import {PropsFor, View} from "@khanacademy/wonder-blocks-core";
 import Icon, {icons} from "@khanacademy/wonder-blocks-icon";
 import {ThemeSwitcherContext, tokens} from "@khanacademy/wonder-blocks-theming";
 import {LabelMedium} from "@khanacademy/wonder-blocks-typography";
@@ -18,19 +18,19 @@ import SwitchArgtypes from "./switch.argtypes";
 type StoryComponentType = StoryObj<typeof Switch>;
 
 /**
- * A Switch is an input that allows users to toggle between two states, typically `on` and `off`.
- * It is a controlled component, meaning that the state of the switch is controlled by the
- * `checked` prop.
- * See the Best Practices tab for more information on how to use this component with labels,
+ * A Switch is an input that allows users to toggle between two states,
+ * typically `on` and `off`. It is a controlled component, meaning that the
+ * state of the switch is controlled by the `checked` prop. See the Best
+ * Practices tab for more information on how to use this component with labels,
  * descriptions, tooltips, and more.
  *
  * ### Usage
  * ```jsx
  *
- * import Switch from "@khanacademy/wonder-blocks-switch";
+ *import Switch from "@khanacademy/wonder-blocks-switch";
  *
- * <Switch checked={false} onChange={() => {}} />
- * ```
+ *<Switch checked={false} onChange={() => {}} />
+ *```
  */
 export default {
     title: "Switch",
@@ -46,11 +46,26 @@ export default {
     argTypes: SwitchArgtypes,
 } as Meta<typeof Switch>;
 
+function ControlledDefaultSwitch(args: PropsFor<typeof Switch>) {
+    const [checked, setChecked] = React.useState(args.checked);
+
+    // Update the checked state when the args change.
+    React.useEffect(() => {
+        setChecked(args.checked);
+    }, [args.checked]);
+
+    return <Switch {...args} checked={checked} onChange={setChecked} />;
+}
+
+/**
+ * The switch has a default state that can be controlled by the `checked` prop.
+ */
 export const Default: StoryComponentType = {
     args: {
         checked: false,
         onChange: () => {},
     },
+    render: ControlledDefaultSwitch,
     decorators: [
         (Story) => (
             <View>
@@ -66,7 +81,6 @@ export const Default: StoryComponentType = {
  * The `onChange` prop is optional in case the toggle will
  * be wrapped in a larger clickable component.
  */
-
 export const Controlled: StoryComponentType = () => {
     const [checkedOne, setCheckedOne] = React.useState(false);
     const [checkedTwo, setCheckedTwo] = React.useState(false);
@@ -110,7 +124,9 @@ Controlled.play = async ({canvasElement}) => {
 };
 
 /**
- * The switch can be disabled.
+ * The switch can be disabled. Note that we use `aria-disabled` to allow the
+ * switch to receive focus even when disabled. This helps Screen Readers to
+ * announce the state of the switch.
  */
 export const Disabled: StoryComponentType = {
     render: () => (