+++ packages/wonder-blocks-dropdown/src/components/multi-select.tsx
@@ -15,7 +15,11 @@ import {
 } from "../util/constants";
 
 import OptionItem from "./option-item";
-import type {DropdownItem, OpenerProps} from "../util/types";
+import type {
+    DropdownItem,
+    OpenerProps,
+    OptionItemComponentArray,
+} from "../util/types";
 
 export type Labels = {
     /**
@@ -64,6 +68,10 @@ type DefaultProps = Readonly<{
      * and does not support interaction. Defaults to false.
      */
     disabled: boolean;
+    /**
+     * Whether this component is in an error state. Defaults to false.
+     */
+    error: boolean;
     /**
      * Whether to display the "light" version of this component instead, for
      * use when the component is used on a dark background.
@@ -202,6 +210,7 @@ export default class MultiSelect extends React.Component<Props, State> {
     static defaultProps: DefaultProps = {
         alignment: "left",
         disabled: false,
+        error: false,
         light: false,
         shortcuts: false,
         selectedValues: [],
@@ -285,11 +294,7 @@ export default class MultiSelect extends React.Component<Props, State> {
         onChange([]);
     };
 
-    getMenuText(
-        children: Array<
-            React.ReactElement<React.ComponentProps<typeof OptionItem>>
-        >,
-    ): string {
+    getMenuText(children: OptionItemComponentArray): string {
         const {implicitAllEnabled, selectedValues} = this.props;
         const {noneSelected, someSelected, allSelected} = this.state.labels;
 
@@ -317,7 +322,7 @@ export default class MultiSelect extends React.Component<Props, State> {
         }
     }
 
-    getShortcuts(numOptions: number): Array<DropdownItem> {
+    getShortcuts(numOptions: number): DropdownItem[] {
         const {selectedValues, shortcuts} = this.props;
         const {selectAllLabel, selectNoneLabel} = this.state.labels;
 
@@ -363,11 +368,7 @@ export default class MultiSelect extends React.Component<Props, State> {
         }
     }
 
-    getMenuItems(
-        children: Array<
-            React.ReactElement<React.ComponentProps<typeof OptionItem>>
-        >,
-    ): Array<DropdownItem> {
+    getMenuItems(children: OptionItemComponentArray): DropdownItem[] {
         const {isFilterable} = this.props;
         // If it's not filterable, no need to do any extra besides mapping the
         // option items to dropdown items.
@@ -386,10 +387,12 @@ export default class MultiSelect extends React.Component<Props, State> {
                 props.label.toLowerCase().indexOf(lowercasedSearchText) > -1,
         );
 
-        // @ts-expect-error [FEI-5019] - TS2315 - Type 'Element' is not generic.
-        const lastSelectedChildren: Array<Element<typeof OptionItem>> = [];
-        // @ts-expect-error [FEI-5019] - TS2315 - Type 'Element' is not generic.
-        const restOfTheChildren: Array<Element<typeof OptionItem>> = [];
+        const lastSelectedChildren: React.ReactElement<
+            React.ComponentProps<typeof OptionItem>
+        >[] = [];
+        const restOfTheChildren: React.ReactElement<
+            React.ComponentProps<typeof OptionItem>
+        >[] = [];
         for (const child of filteredChildren) {
             if (lastSelectedValues.includes(child.props.value)) {
                 lastSelectedChildren.push(child);
@@ -454,9 +457,9 @@ export default class MultiSelect extends React.Component<Props, State> {
     };
 
     renderOpener(
-        allChildren: Array<
-            React.ReactElement<React.ComponentProps<typeof OptionItem>>
-        >,
+        allChildren: React.ReactElement<
+            React.ComponentProps<typeof OptionItem>
+        >[],
     ):
         | React.ReactElement<React.ComponentProps<typeof DropdownOpener>>
         | React.ReactElement<React.ComponentProps<typeof SelectOpener>> {
@@ -481,6 +484,8 @@ export default class MultiSelect extends React.Component<Props, State> {
             shortcuts,
             style,
             className,
+            "aria-invalid": ariaInvalid,
+            "aria-required": ariaRequired,
             /* eslint-enable @typescript-eslint/no-unused-vars */
             ...sharedProps
         } = this.props;
@@ -526,16 +531,20 @@ export default class MultiSelect extends React.Component<Props, State> {
             dropdownStyle,
             children,
             isFilterable,
+            "aria-invalid": ariaInvalid,
+            "aria-required": ariaRequired,
         } = this.props;
         const {open, searchText} = this.state;
         const {clearSearch, filter, noResults, someSelected} =
             this.state.labels;
 
-        const allChildren = React.Children.toArray(children).filter(Boolean);
+        const allChildren = (
+            React.Children.toArray(children) as Array<
+                React.ReactElement<React.ComponentProps<typeof OptionItem>>
+            >
+        ).filter(Boolean);
         const numOptions = allChildren.length;
-        // @ts-expect-error [FEI-5019] - TS2345 - Argument of type '(ReactChild | ReactFragment | ReactPortal)[]' is not assignable to parameter of type 'ReactElement<{}, string | JSXElementConstructor<any>>[]'.
         const filteredItems = this.getMenuItems(allChildren);
-        // @ts-expect-error [FEI-5019] - TS2345 - Argument of type '(ReactChild | ReactFragment | ReactPortal)[]' is not assignable to parameter of type 'ReactElement<{}, string | JSXElementConstructor<any>>[]'.
         const opener = this.renderOpener(allChildren);
 
         return (
@@ -567,6 +576,8 @@ export default class MultiSelect extends React.Component<Props, State> {
                     noResults,
                     someResults: someSelected,
                 }}
+                aria-invalid={ariaInvalid}
+                aria-required={ariaRequired}
             />
         );
     }