+++ packages/wonder-blocks-icon/src/index.ts
@@ -2,6 +2,11 @@ import Icon from "./components/icon";
 import type {IconAsset, IconSize} from "./util/icon-assets";
 
 export * as icons from "./util/icon-assets";
-
+export {PhosphorIcon} from "./components/phosphor-icon";
+export type {
+    PhosphorIconAsset,
+    PhosphorIconMedium,
+    PhosphorIconSmall,
+} from "./types";
 export type {IconAsset, IconSize};
 export default Icon;