+++ packages/wonder-blocks-icon/src/types.ts
@@ -1,0 +1,12 @@
+/**
+ * All the possible icon weights.
+ */
+export type PhosphorIconAsset = PhosphorRegular | PhosphorBold | PhosphorFill;
+/**
+ * The different icon weights for small icons.
+ */
+export type PhosphorIconSmall = PhosphorBold | PhosphorFill;
+/**
+ * The different icon weights for medium icons.
+ */
+export type PhosphorIconMedium = PhosphorRegular | PhosphorFill;