+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 0.0.5
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-clickable@4.0.8
+
+## 0.0.3
+
+### Patch Changes
+
+-   Updated dependencies [80cab317]
+    -   @khanacademy/wonder-blocks-clickable@4.0.7
+
 ## 0.0.2
 
 ### Patch Changes