+++ packages/wonder-blocks-banner/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @khanacademy/wonder-blocks-banner
 
+## 3.0.22
+
+### Patch Changes
+
+-   cc6b1950: Switch internal `IconButton` instances to use phosphor.
+-   Updated dependencies [cc6b1950]
+    -   @khanacademy/wonder-blocks-icon-button@5.0.0
+
+## 3.0.21
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+    -   @khanacademy/wonder-blocks-button@4.2.5
+    -   @khanacademy/wonder-blocks-icon-button@4.2.2
+    -   @khanacademy/wonder-blocks-link@5.0.9
+
+## 3.0.20
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-button@4.2.4
+    -   @khanacademy/wonder-blocks-icon-button@4.2.1
+    -   @khanacademy/wonder-blocks-link@5.0.8
+
+## 3.0.19
+
+### Patch Changes
+
+-   Updated dependencies [fb704043]
+    -   @khanacademy/wonder-blocks-icon-button@4.2.0
+    -   @khanacademy/wonder-blocks-button@4.2.3
+    -   @khanacademy/wonder-blocks-link@5.0.7
+
 ## 3.0.18
 
 ### Patch Changes