+++ packages/wonder-blocks-button/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @khanacademy/wonder-blocks-button
 
+## 4.2.5
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+
+## 4.2.4
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-clickable@4.0.8
+    -   @khanacademy/wonder-blocks-progress-spinner@2.0.23
+
+## 4.2.3
+
+### Patch Changes
+
+-   Updated dependencies [80cab317]
+    -   @khanacademy/wonder-blocks-clickable@4.0.7
+
 ## 4.2.2
 
 ### Patch Changes