+++ packages/wonder-blocks-labeled-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-labeled-field
 
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [b6fbd635]
+    -   @khanacademy/wonder-blocks-layout@2.0.23
+
+## 0.1.3
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+
 ## 0.1.2
 
 ### Patch Changes