+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-link
 
+## 5.0.9
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+
+## 5.0.8
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-clickable@4.0.8
+
+## 5.0.7
+
+### Patch Changes
+
+-   Updated dependencies [80cab317]
+    -   @khanacademy/wonder-blocks-clickable@4.0.7
+
 ## 5.0.6
 
 ### Patch Changes