+++ packages/wonder-blocks-modal/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @khanacademy/wonder-blocks-modal
 
+## 4.0.30
+
+### Patch Changes
+
+-   cc6b1950: Switch internal `IconButton` instances to use phosphor.
+-   Updated dependencies [b6fbd635]
+-   Updated dependencies [cc6b1950]
+    -   @khanacademy/wonder-blocks-layout@2.0.23
+    -   @khanacademy/wonder-blocks-icon-button@5.0.0
+
+## 4.0.29
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+    -   @khanacademy/wonder-blocks-icon-button@4.2.2
+
+## 4.0.28
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-icon-button@4.2.1
+
+## 4.0.27
+
+### Patch Changes
+
+-   Updated dependencies [fb704043]
+    -   @khanacademy/wonder-blocks-icon-button@4.2.0
+
 ## 4.0.26
 
 ### Patch Changes