+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 2.1.17
+
+### Patch Changes
+
+-   cc6b1950: Switch internal `IconButton` instances to use phosphor.
+-   Updated dependencies [cc6b1950]
+    -   @khanacademy/wonder-blocks-icon-button@5.0.0
+    -   @khanacademy/wonder-blocks-form@4.3.13
+
+## 2.1.16
+
+### Patch Changes
+
+-   Updated dependencies [ea0e7c02]
+    -   @khanacademy/wonder-blocks-icon@2.2.0
+    -   @khanacademy/wonder-blocks-form@4.3.12
+    -   @khanacademy/wonder-blocks-icon-button@4.2.2
+
+## 2.1.15
+
+### Patch Changes
+
+-   Updated dependencies [48d3c7e9]
+    -   @khanacademy/wonder-blocks-color@3.0.0
+    -   @khanacademy/wonder-blocks-form@4.3.11
+    -   @khanacademy/wonder-blocks-icon-button@4.2.1
+
+## 2.1.14
+
+### Patch Changes
+
+-   Updated dependencies [fb704043]
+    -   @khanacademy/wonder-blocks-icon-button@4.2.0
+    -   @khanacademy/wonder-blocks-form@4.3.10
+
 ## 2.1.13
 
 ### Patch Changes